Remove `ResultOk` to Utils
